Lambek's theorem & proof
This requires an additional property of EndoFunctors: mapping of the morphisms must preserve morphism equality, aka `forall x y : A → B, x ≡ y -> EF_mor x ≡ EF_mor y`. I am not sure if this property can be proved from the previous ones, so I added it as a law for now.

This code could be cleaned up a bit.